refatcor: move smoother constructor definitions back to cpp
