+++ planning/autoware_velocity_smoother/include/autoware/velocity_smoother/smoother/jerk_filtered_smoother.hpp
@@ -45,18 +45,7 @@ public:
 
   template <typename NodeT>
   explicit JerkFilteredSmoother(
-    NodeT & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper)
-  : SmootherBase(node, time_keeper)
-  {
-    auto & p = smoother_param_;
-    p.jerk_weight = node.template declare_parameter<double>("jerk_weight");
-    p.over_v_weight = node.template declare_parameter<double>("over_v_weight");
-    p.over_a_weight = node.template declare_parameter<double>("over_a_weight");
-    p.over_j_weight = node.template declare_parameter<double>("over_j_weight");
-    p.jerk_filter_ds = node.template declare_parameter<double>("jerk_filter_ds");
-
-    initQpInterface();
-  }
+    NodeT & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
 
   bool apply(
     const double initial_vel, const double initial_acc, const TrajectoryPoints & input,
@@ -75,8 +64,6 @@ private:
   Param smoother_param_;
   std::shared_ptr<autoware::qp_interface::QPInterface> qp_interface_;
   rclcpp::Logger logger_{rclcpp::get_logger("smoother").get_child("jerk_filtered_smoother")};
-
-  void initQpInterface();
 
   TrajectoryPoints forwardJerkFilter(
     const double v0, const double a0, const double a_max, const double a_stop, const double j_max,
