+++ planning/autoware_velocity_smoother/src/smoother/jerk_filtered_smoother.cpp
@@ -18,6 +18,7 @@
 #include "autoware/velocity_smoother/trajectory_utils.hpp"
 
 #include <Eigen/Core>
+#include <autoware/agnocast_wrapper/node.hpp>
 
 #include <algorithm>
 #include <chrono>
@@ -32,11 +33,27 @@
 
 namespace autoware::velocity_smoother
 {
-void JerkFilteredSmoother::initQpInterface()
+template <typename NodeT>
+JerkFilteredSmoother::JerkFilteredSmoother(
+  NodeT & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper)
+: SmootherBase(node, time_keeper)
 {
+  auto & p = smoother_param_;
+  p.jerk_weight = node.template declare_parameter<double>("jerk_weight");
+  p.over_v_weight = node.template declare_parameter<double>("over_v_weight");
+  p.over_a_weight = node.template declare_parameter<double>("over_a_weight");
+  p.over_j_weight = node.template declare_parameter<double>("over_j_weight");
+  p.jerk_filter_ds = node.template declare_parameter<double>("jerk_filter_ds");
+
   qp_interface_ =
     std::make_shared<autoware::qp_interface::ProxQPInterface>(false, 20000, 1.0e-8, 1.0e-6, false);
 }
+
+template JerkFilteredSmoother::JerkFilteredSmoother(
+  rclcpp::Node & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
+template JerkFilteredSmoother::JerkFilteredSmoother(
+  autoware::agnocast_wrapper::Node & node,
+  const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
 
 void JerkFilteredSmoother::setParam(const Param & smoother_param)
 {
