+++ planning/autoware_velocity_smoother/src/smoother/linf_pseudo_jerk_smoother.cpp
@@ -17,6 +17,7 @@
 #include "autoware/velocity_smoother/trajectory_utils.hpp"
 
 #include <Eigen/Core>
+#include <autoware/agnocast_wrapper/node.hpp>
 
 #include <algorithm>
 #include <chrono>
@@ -26,6 +27,28 @@
 
 namespace autoware::velocity_smoother
 {
+template <typename NodeT>
+LinfPseudoJerkSmoother::LinfPseudoJerkSmoother(
+  NodeT & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper)
+: SmootherBase(node, time_keeper)
+{
+  auto & p = smoother_param_;
+  p.pseudo_jerk_weight = node.template declare_parameter<double>("pseudo_jerk_weight");
+  p.over_v_weight = node.template declare_parameter<double>("over_v_weight");
+  p.over_a_weight = node.template declare_parameter<double>("over_a_weight");
+
+  qp_solver_.updateMaxIter(20000);
+  qp_solver_.updateRhoInterval(5000);
+  qp_solver_.updateEpsRel(1.0e-4);  // def: 1.0e-4
+  qp_solver_.updateEpsAbs(1.0e-8);  // def: 1.0e-4
+  qp_solver_.updateVerbose(false);
+}
+
+template LinfPseudoJerkSmoother::LinfPseudoJerkSmoother(
+  rclcpp::Node & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
+template LinfPseudoJerkSmoother::LinfPseudoJerkSmoother(
+  autoware::agnocast_wrapper::Node & node,
+  const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
 
 void LinfPseudoJerkSmoother::setParam(const Param & smoother_param)
 {
