+++ planning/autoware_velocity_smoother/src/smoother/smoother_base.cpp
@@ -20,6 +20,7 @@
 #include "autoware/velocity_smoother/resample.hpp"
 #include "autoware/velocity_smoother/trajectory_utils.hpp"
 
+#include <autoware/agnocast_wrapper/node.hpp>
 #include <autoware_utils_geometry/geometry.hpp>
 #include <autoware_utils_math/unit_conversion.hpp>
 
@@ -63,6 +64,54 @@ TrajectoryPoints applyPreProcess(
   return output;
 }
 }  // namespace
+
+template <typename NodeT>
+SmootherBase::SmootherBase(
+  NodeT & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper)
+: time_keeper_(time_keeper)
+{
+  auto & p = base_param_;
+  p.max_accel = node.template declare_parameter<double>("normal.max_acc");
+  p.min_decel = node.template declare_parameter<double>("normal.min_acc");
+  p.stop_decel = node.template declare_parameter<double>("stop_decel");
+  p.max_jerk = node.template declare_parameter<double>("normal.max_jerk");
+  p.min_jerk = node.template declare_parameter<double>("normal.min_jerk");
+  p.min_decel_for_lateral_acc_lim_filter =
+    node.template declare_parameter<double>("min_decel_for_lateral_acc_lim_filter");
+  p.sample_ds = node.template declare_parameter<double>("resample_ds");
+  p.curvature_threshold = node.template declare_parameter<double>("curvature_threshold");
+  p.lateral_acceleration_limits =
+    node.template declare_parameter<std::vector<double>>("lateral_acceleration_limits");
+  p.velocity_thresholds =
+    node.template declare_parameter<std::vector<double>>("velocity_thresholds");
+  p.steering_angle_rate_limits =
+    node.template declare_parameter<std::vector<double>>("steering_angle_rate_limits");
+  p.curvature_calculation_distance =
+    node.template declare_parameter<double>("curvature_calculation_distance");
+  p.decel_distance_before_curve =
+    node.template declare_parameter<double>("decel_distance_before_curve");
+  p.decel_distance_after_curve =
+    node.template declare_parameter<double>("decel_distance_after_curve");
+  p.min_curve_velocity = node.template declare_parameter<double>("min_curve_velocity");
+  p.resample_param.max_trajectory_length =
+    node.template declare_parameter<double>("max_trajectory_length");
+  p.resample_param.min_trajectory_length =
+    node.template declare_parameter<double>("min_trajectory_length");
+  p.resample_param.resample_time = node.template declare_parameter<double>("resample_time");
+  p.resample_param.dense_resample_dt = node.template declare_parameter<double>("dense_resample_dt");
+  p.resample_param.dense_min_interval_distance =
+    node.template declare_parameter<double>("dense_min_interval_distance");
+  p.resample_param.sparse_resample_dt =
+    node.template declare_parameter<double>("sparse_resample_dt");
+  p.resample_param.sparse_min_interval_distance =
+    node.template declare_parameter<double>("sparse_min_interval_distance");
+}
+
+template SmootherBase::SmootherBase(
+  rclcpp::Node & node, const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
+template SmootherBase::SmootherBase(
+  autoware::agnocast_wrapper::Node & node,
+  const std::shared_ptr<autoware_utils_debug::TimeKeeper> time_keeper);
 
 void SmootherBase::setWheelBase(const double wheel_base)
 {
